:recycle: Be independent from the git client
